+++ client/auth-token.js
