+++ lib/api-cleanup.js
